[bndtools.core] Moved launch delegates into api package
